Fix slow-charge divider rounding, factory reset, and log replay
Slow charge (CANBUS.cpp):
- The taper truncated twice - integer division to mA, then again to whole
  deciamps in the CAN senders - so it always rounded down. The UI's own
  recommended divider of 24 on a 280Ah pack asked for 11666 mA and went out
  as 11.6A. Round the division, then round to a whole 0.1A.
- Clamp the taper to the configured maximum. Capacity/divider knows nothing
  of that ceiling, so a big pack on a small divider landed above it and a
  slow charge could end up raising the limit.
- 0x35F capacity went via whole Ah before scaling, reporting a 280.5Ah pack
  as 280.0.

Factory reset (mEEPROM.cpp):
- Preferences::clear() only empties the namespace its handle is open on.
  WiFi and MQTT settings live in the separate "network" namespace, so
  Factory Reset left credentials behind entirely, and Keep WiFi only kept
  them by accident. Erase the whole NVS partition instead.
- The keep-WiFi path wrote the SSID twice and never the password, called
  begin() on an already-open handle (which returns false, so the restore
  never ran at all), and returned an uninitialised bool.
- Battery capacity was written with putUInt8 on first boot while every read
  uses getUInt32.

Log viewer (HTTPWSFunctions.h, index.htm.template):
- GetLogs() walked forward from logBufferIndex, handing over the oldest 50
  of the 100 entry ring rather than the newest 50, so the tab showed the
  boot sequence and nothing since.
- Replayed lines carry an age now. The browser stamped every line with its
  arrival time, collapsing a six minute backlog onto one second and making
  it read as a reboot that never happened.
- Pull the backlog once per connection, not on every visit to the tab.
- Share the JSON escaping between the live and replay paths; the replay had
  none, and built into a 256 byte buffer that a full length message could
  truncate mid-string. Client tolerates a malformed frame instead of losing
  the message handler to it.

+++ src/main.cpp
@@ -130,7 +130,7 @@ void setup()
     pref.putUInt8(ccSlowSOCCharge2,0);
     pref.putUInt8(ccSlowSOCDivider1,0);
     pref.putUInt8(ccSlowSOCDivider2,0);
-    pref.putUInt8(ccBattCapacity, initBattCapacity);
+    pref.putUInt32(ccBattCapacity, initBattCapacity);
     pref.putBool(ccPylonTech, false);
     pref.putBool(ccSOCTrick, false);
     pref.putBool(ccRequestFlags, false);
